Changed modules dependencies, moved spring configuration file to Web module
