+++ persistence/build.gradle
@@ -13,7 +13,6 @@ dependencies {
     compile project(":model")
     compile group: 'org.springframework', name: 'spring-context', version: '5.2.9.RELEASE'
     compile group: 'org.springframework', name: 'spring-jdbc', version: '5.2.9.RELEASE'
-    compile group: 'org.springframework', name: 'spring-core', version: '5.2.9.RELEASE'
     compile group: 'mysql', name: 'mysql-connector-java', version: '8.0.22'
     testCompile group: 'junit', name: 'junit', version: '4.12'
 }
