--- persistence/src/main/java/com/epam/esm/config/AppConfig.java
@@ -1,36 +1,0 @@
-package com.epam.esm.config;
-
-import com.epam.esm.dao.impl.SQLDaoTagImpl;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.datasource.DriverManagerDataSource;
-
-import javax.sql.DataSource;
-
-@Configuration
-@ComponentScan("com.epam.esm")
-public class AppConfig {
-
-    @Bean
-    public DataSource dataSource() {
-        DriverManagerDataSource dataSource = new DriverManagerDataSource();
-        dataSource.setDriverClassName("com.mysql.jdbc.Driver");
-        dataSource.setUrl("jdbc:mysql://localhost:3306/GiftShop");
-        dataSource.setUsername("root");
-        dataSource.setPassword("12345678");
-
-        return dataSource;
-    }
-
-    @Bean
-    public JdbcTemplate jdbcTemplate() {
-        return new JdbcTemplate(dataSource());
-    }
-
-    @Bean
-    public SQLDaoTagImpl tagDAO() {
-        return new SQLDaoTagImpl();
-    }
-}
