+++ web/build.gradle
@@ -11,6 +11,11 @@ repositories {
 }
 
 dependencies {
+    compile project(":persistence")
+    compile group: 'org.springframework', name: 'spring-webmvc', version: '5.2.9.RELEASE'
+    compile group: 'org.springframework', name: 'spring-jdbc', version: '5.2.9.RELEASE'
+    compile group: 'org.springframework', name: 'spring-context', version: '5.2.9.RELEASE'
+    compile group: 'org.springframework', name: 'spring-core', version: '5.2.9.RELEASE'
     testCompile group: 'junit', name: 'junit', version: '4.11'
     testCompile group: 'junit', name: 'junit', version: '4.12'
 }
